@@ -1,5 +1,5 @@
 ﻿using System.Text.Json;
-using JiraRestClient.Net.Core.Converter;
+using JiraRestClient.Net.Core.Extension;
 using JiraRestClient.Net.Domain;
 using JiraRestClient.Net.Util;
 
@@ -11,33 +11,33 @@ namespace JiraRestClient.Net.Core
         {
         }
 
-        public Permissions GetMyPermissions(string projectKey = null, 
-            string projectId = null, 
+        public Permissions GetMyPermissions(string projectKey = null,
+            string projectId = null,
             string issueKey = null,
             string issueId = null)
         {
-            var restUriBuilder = UriHelper.BuildPath(BaseUri, RestPathConstants.MyPermissions);
+            var uri = BaseUri.AddPaths(RestPathConstants.MyPermissions);
             if (string.IsNullOrEmpty(projectKey) == false)
             {
-                UriHelper.AddQuery(restUriBuilder, RestParamConstants.ProjectKey, projectKey);
+                uri = uri.AddQuery($"{RestParamConstants.ProjectKey}={projectKey}");
             }
 
             if (string.IsNullOrEmpty(projectId) == false)
             {
-                UriHelper.AddQuery(restUriBuilder, RestParamConstants.ProjectId, projectId);
+                uri = uri.AddQuery($"{RestParamConstants.ProjectId}={projectId}");
             }
 
             if (string.IsNullOrEmpty(issueKey) == false)
             {
-                UriHelper.AddQuery(restUriBuilder, RestParamConstants.IssueKey, issueKey);
+                uri = uri.AddQuery($"{RestParamConstants.IssueKey}={issueKey}");
             }
 
             if (string.IsNullOrEmpty(issueId) == false)
             {
-                UriHelper.AddQuery(restUriBuilder, RestParamConstants.IssueId, issueId);
+                uri = uri.AddQuery($"{RestParamConstants.IssueId}={issueId}");
             }
 
-            var stream = Client.GetStringAsync(restUriBuilder.ToString());
+            var stream = Client.GetStringAsync(uri.ToString());
             var streamResult = stream.Result;
             return JsonSerializer.Deserialize<Permissions>(streamResult);
         }
