@@ -1,5 +1,6 @@
 using System.Collections.Generic;
 using System.Text.Json;
+using JiraRestClient.Net.Core.Extension;
 using JiraRestClient.Net.Domain;
 using JiraRestClient.Net.Util;
 
@@ -16,7 +17,7 @@ namespace JiraRestClient.Net.Core
 
         public Project GetProjectByKey(string key)
         {
-            var restUriBuilder = UriHelper.BuildPath(BaseUri, RestPathConstants.Project, key);
+            var restUriBuilder = BaseUri.AddPaths(RestPathConstants.Project, key);
             var completeUri = restUriBuilder.ToString();
             var stream = Client.GetStringAsync(completeUri);
             return JsonSerializer.Deserialize<Project>(stream.Result);
@@ -24,19 +25,19 @@ namespace JiraRestClient.Net.Core
 
         public IEnumerable<Project> GetAllProjects()
         {
-            var restUriBuilder = UriHelper.BuildPath(BaseUri, RestPathConstants.Project);
+            var restUriBuilder = BaseUri.AddPaths(RestPathConstants.Project);
             var stream = Client.GetStringAsync(restUriBuilder.ToString());
             return JsonSerializer.Deserialize<List<Project>>(stream.Result);
         }
 
         public List<Version> GetProjectVersions(string key){
-            var restUriBuilder = UriHelper.BuildPath(BaseUri, RestPathConstants.Project, key, RestPathConstants.Versions);
+            var restUriBuilder = BaseUri.AddPaths(RestPathConstants.Project, key, RestPathConstants.Versions);
             var stream = Client.GetStringAsync(restUriBuilder.ToString());
             return JsonSerializer.Deserialize<List<Version>>(stream.Result);
         }
 
         public List<Component> GetProjectComponents(string key){
-            var restUriBuilder = UriHelper.BuildPath(BaseUri, RestPathConstants.Project, key, RestPathConstants.Components);
+            var restUriBuilder = BaseUri.AddPaths(RestPathConstants.Project, key, RestPathConstants.Components);
             var stream = Client.GetStringAsync(restUriBuilder.ToString());
             return JsonSerializer.Deserialize<List<Component>>(stream.Result);
         }
