@@ -1,6 +1,7 @@
 using System.Net.Http;
 using System.Text;
 using System.Text.Json;
+using JiraRestClient.Net.Core.Extension;
 using JiraRestClient.Net.Jql;
 using JiraRestClient.Net.Util;
 
@@ -15,7 +16,7 @@ namespace JiraRestClient.Net.Core
         public JqlSearchResult SearchIssues(JqlSearchBean jqlSearchBean)
         {
             var json = JsonSerializer.Serialize(jqlSearchBean);
-            var uri = UriHelper.BuildPath(BaseUri, RestPathConstants.Search);
+            var uri = BaseUri.AddPaths(RestPathConstants.Search);
             var httpContent = new StringContent(json, Encoding.UTF8, "application/json");
             var response = Client.PostAsync(uri.ToString(), httpContent);
             var readAsStringAsync = response.Result.Content.ReadAsStringAsync();
